More options screen ready for Navigation compose
